+++ nstep_replay_mem_prioritized.pyx
@@ -43,8 +43,8 @@ cdef class py_Data:
         node_feats = graph1.node_feats
         EdgeWeight = graph1.EdgeWeight
 
-        cint_edges_from = np.zeros([num_edges],dtype=np.int)
-        cint_edges_to = np.zeros([num_edges],dtype=np.int)
+        cint_edges_from = np.zeros([num_edges],dtype=np.int32)
+        cint_edges_to = np.zeros([num_edges],dtype=np.int32)
         cdouble_vec_node_feats = np.zeros([num_nodes, 2], dtype=np.double)
         cdouble_EdgeWeight = np.zeros([num_nodes, num_nodes], dtype=np.double)
         # print(num_nodes)
@@ -182,8 +182,8 @@ cdef class py_ReplaySample:
         node_feats = graph1.node_feats
         EdgeWeight = graph1.EdgeWeight
 
-        cint_edges_from = np.zeros([num_edges],dtype=np.int)
-        cint_edges_to = np.zeros([num_edges],dtype=np.int)
+        cint_edges_from = np.zeros([num_edges],dtype=np.int32)
+        cint_edges_to = np.zeros([num_edges],dtype=np.int32)
         cdouble_vec_node_feats = np.zeros([num_nodes, 2], dtype=np.double)
         cdouble_EdgeWeight = np.zeros([num_nodes, num_nodes], dtype=np.double)
         # print(num_nodes)
